tea-disease-detection/requirements.txt
Flask==3.0.0
tensorflow==2.15.0
Pillow==10.1.0
numpy==1.24.3
werkzeug==3.0.1
